Refactor: Extract separation of concerns
This introduces the separation of application-specific logic (i.e. how
exactly we want to merge two things together, the binary operator of the
monoid) from the application-agnostic infrastructure to get it done via
Sidekiq in a concurrent way.

This makes the monoid act on Filepaths, albeit with a little funkiness
to make use of stdlib temporary files.